+++ .gitignore
@@ -35,7 +35,7 @@ MANIFEST
 
 # PyInstaller
 #  Usually these files are written by a python script from a template
-#  before PyInstaller builds the exe, so as to inject date/other infos into it.
+#  before PyInstaller builds the exe, so as to inject date/pipeline infos into it.
 *.manifest
 *.spec
