+++ pipeline/config/constants.py
@@ -1,0 +1,1 @@
+TARGET = "median_house_value"
