+++ pipeline/evaluate.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+"""
+Date: 02/2023
+Version: 1.0
+Author: (I) Jose Pena
+Website: https://github.com/JoseJuan98
+
+
+Title
+=====
+
+...
+"""
