--- src/exper/utils/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Script to include all scripts in this folder in the namespace of utils."""
-from src.pipeline.utils.pipeline import pipe_args, parse_args, create_summary
-from src.pipeline.utils.data_transformations import get_preprocessor
