--- src/other/constants.py
@@ -1,1 +1,0 @@
-TARGET = "median_house_value"
